Use asExternalUri For Server Url
This allows for automatic port forwarding of trace-server api calls when using remote.

Signed-off-by: Will Yang <william.yang@ericsson.com>

Server offline landing view

Creates a new webview in the trace explorer.  This simply displays a button that allows
the user to start the server if the front-end detects the server is offline.  This:

1. Provides a more intuitive user experience
2. Stops the user from sendinding some requests when the server is not online which
      causes the application to break

Changes here are tied to changes in the theia-trace-extension repository.

Signed-off-by: William Yang <william.yang@ericsson.com>